+++ src/app/routes/connector-ui/connector-ui.component.ts
@@ -4,6 +4,7 @@ import {Title} from '@angular/platform-browser';
 import {Observable} from 'rxjs';
 import {map, shareReplay} from 'rxjs/operators';
 import {APP_CONFIG, AppConfig} from '../../core/config/app-config';
+import {LoginPollingService} from '../../core/services/login-polling.service';
 import {TitleUtilsService} from '../../core/services/title-utils.service';
 import {routes} from './connector-ui-routing.module';
 
@@ -14,6 +15,7 @@ import {routes} from './connector-ui-routing.module';
   providers: [TitleUtilsService],
 })
 export class ConnectorUiComponent implements OnInit {
+
   isHandset$: Observable<boolean> = this.breakpointObserver
     .observe(Breakpoints.Handset)
     .pipe(
@@ -28,9 +30,15 @@ export class ConnectorUiComponent implements OnInit {
     public titleUtilsService: TitleUtilsService,
     public titleService: Title,
     private breakpointObserver: BreakpointObserver,
+    private loginPollingService: LoginPollingService,
   ) {}
 
   ngOnInit() {
     this.titleUtilsService.startUpdatingTitleFromRouteData('EDC Connector');
+    this.startLoginPolling();
+  }
+
+  private startLoginPolling() {
+    this.loginPollingService.startPolling();
   }
 }